30/9 new

+++ src/main/java/com/example/hiepphat/repositories/RecipeRepository.java
@@ -25,6 +25,6 @@ public interface RecipeRepository extends JpaRepository<Recipe,Long> {
     int countByUser_UserID(int userID);
     List<Recipe> findByRecipeTitleLike(String search);
     @Query(value = "Select recipe_id from Recipes where recipe_title=? and user_id=?",nativeQuery = true)
-    int findrecipeID(String title,int userID);
+    long findrecipeID(String title,int userID);
 
 }
